feat: 스토리북 preview 확장자 tsx로 변경
