--- fe/.storybook/preview.ts
@@ -1,22 +1,0 @@
-import { GlobalStyles } from './../src/styles/globalStyles';
-import type { Preview } from '@storybook/react';
-import { withThemeFromJSXProvider } from '@storybook/addon-styling';
-
-const preview: Preview = {
-  parameters: {
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/i,
-      },
-    },
-  },
-};
-
-export const decorators = [
-  withThemeFromJSXProvider({
-    GlobalStyles,
-  }),
-];
-
-export default preview;
